Final Bank convert float to long, closes #130 (#146)
* Final Bank convert float to long, closes #130

* Fixes Bank convert float to long, closes #130

* Remove imports from Bank convert float to long, closes#130

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Account.java
@@ -36,7 +36,7 @@ public class Account extends Account_Base {
 
 	}
 
-	public Operation deposit(double amount) {
+	public Operation deposit(long amount) {
 		if (amount <= 0) {
 			throw new BankException();
 		}
@@ -46,7 +46,7 @@ public class Account extends Account_Base {
 		return new Operation(Operation.Type.DEPOSIT, this, amount);
 	}
 
-	public Operation withdraw(double amount) {
+	public Operation withdraw(long amount) {
 		if (amount <= 0 || amount > getBalance()) {
 			throw new BankException();
 		}
